Fix video quality regression CI issues
Addresses review and CI feedback by hardening workflow inputs against shell
injection, moving video-quality implementation into the installable
flashdreams.quality package, keeping tools.video_quality as compatibility
wrappers, tightening manifest validation, and making known-bad calibration
fail on unexpected threshold failures.
