+++ flashdreams/flashdreams/quality/video_quality/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Compatibility package for repository-local video-quality tooling."""
+"""Manifest-driven video-quality regression tooling."""
